requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==6.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.2.0
psycopg[binary]==3.2.12
beautifulsoup4==4.12.3
requests==2.31.0
